Post Comp 1 (#33)
* Hella updates

* Finally controllr working

* mamafufu

* Drivebase too

* telemetry

* Finally working scoring system... Still needs tweaks

* Specimen Grabber first stab

* Auto

* Comp 1

* Post Qual 1

---------

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/settings/GamepadSettings.java
@@ -6,5 +6,5 @@ public class GamepadSettings {
     public static final double GP1_STICK_DEADZONE = 0.05;
     public static final double GP1_TRIGGER_DEADZONE = 0.25;
 
-    public static final double EXPONENT_MODIFIER = 2;
+    public static final int EXPONENT_MODIFIER = 2;
 }
